### 2020-05-22 #2216 @danielskatz
👋 @blthayer - thanks for your submission to JOSS. Could you take a look a [this list of potential reviewers](https://bit.ly/joss-reviewers) and identify a few people who would be good candidates to review this submission?

Also, be aware that this might be difficult.  I don't know anything about "PowerWorld Simulator Automation Server (SimAuto)", but if this is something reviewers need to know, need to have, or even just need to install, this will limit the number of people who may be able or willing to review.
