### 2020-05-22 #2216 @blthayer
@danielskatz - Thank you for volunteering to be our editor!

I've reviewed the list of potential reviewers, and you're right, there don't seem to be too many in the power systems domain. I did find `robinroche` who listed Python as a language of choice, and Power Systems as a domain area.

Myself and my coauthors planned to reach out into our research community to find individuals who have a PowerWorld license and might be interested in reviewing after we made it to this stage. We have reviewed the [ethics guidelines](https://joss.theoj.org/about#ethics) and the [conflict of interest policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) and will ensure our suggested reviewers meet the guidelines/policies.

I'll kick off our search, and report back as soon as possible with potential reviewers.
