### 2020-05-22 #2216 @danielskatz
👋 @robinroche - Would you be able to review this submission for JOSS?  Given the Covid-19 situation, we are being very flexible on reviewing times
