### 2020-05-25 #2216 @blthayer
@robinroche - thank you for being so responsive. We have primarily done our work/testing with versions 20 and 21, though I know PowerWorld is very dedicated to maintaining backward compatibility in SimAuto.

This is an area we should upgrade ESA's documentation - we have not extensively tested with older versions of PWS/SimAuto.

I suspect if we run into issues, it'll be related to the `.pwb` cases in our repository being saved in the version 20 or 21 format. I believe that we can use a newer version (e.g. version 21) to save files in the version 17 format, so please let us know if we need to do that.

I'm still waiting on some reviewer leads, but I'm hoping we can get this moving this week. In the meantime, I'll try to find some time this week to install version 17 and do some preliminary tests to preempt PWS version issues.

Thanks again!
