### 2020-05-28 3be43c9
1. Allow the get_power_flow_results to have additional fields.
