+++ esa/saw.py
@@ -422,13 +422,17 @@ class SAW(object):
         # Return a listing of the internal field name.
         return key_field_df['internal_field_name'].tolist()
 
-    def get_power_flow_results(self, ObjectType: str) -> \
+    def get_power_flow_results(self, ObjectType: str, OptionalFields: Union[
+        None, List[str]] = None) -> \
             Union[None, pd.DataFrame]:
         """Get the power flow results from SimAuto server.
 
         :param ObjectType: Object type to get results for. Valid types
             are the keys in the POWER_FLOW_FIELDS class attribute (case
             insensitive).
+
+        :param OptionalFields: Pass a list of field names if the default
+            fields are not enough.
 
         :returns: Pandas DataFrame with the corresponding results, or
             None if the given ObjectType is not present in the model.
@@ -439,6 +443,8 @@ class SAW(object):
         # Get the listing of fields for this object type.
         try:
             field_list = self.POWER_FLOW_FIELDS[object_type]
+            if OptionalFields:
+                field_list += OptionalFields
         except KeyError:
             raise ValueError('Unsupported ObjectType for power flow results, '
                              '{}.'.format(ObjectType))
