### 2020-05-28 4b71c98
Adding helper method to get the Simulator version.
