### 2020-05-29 e334690
The UIVisible attribute was not introduced until Simulator version 20. Throwing a warning if the attribute fails to set.
