### 2020-05-29 7d8ca10
Older versions of Simulator (e.g. version 17) do not return a display name with GetFieldList. Fixing SAW and associated tests.
