### 2020-05-29 25dcf3a
Fixing some tests that were broken with Simulator 17 due to a lack of Simulator backwards compatibility.
