### 2020-05-29 fa769a3
Skipping OpenOneLineTestCase and CloseOnelineTestCase if Simulator version is < 21. Later, we should save all .pwd files with different versions and retest.
