### 2020-05-29 b2a492e
The weighted_graph snippet does not work for Simulator 17. Adding a note to the documentation, and skipping it for Simulator versions less than 21.
