### 2020-05-29 5c7e3ef
Updating snippets now that test cases have been suffixed with the Simulator version.
