+++ docs/rst/snippets/increase_loading_14.rst
@@ -3,7 +3,7 @@ If you want to follow along, you'll first need to define your own
 
 .. code:: python
 
-    CASE_PATH = r"C:\Users\myuser\git\ESA\tests\cases\ieee_14\IEEE 14 bus.pwb"
+    CASE_PATH = r"C:\Users\myuser\git\ESA\tests\cases\ieee_14\IEEE 14 bus_pws_version_21.pwb"
 
 Then, import the SimAuto wrapper (SAW) class and initialize an instance:
 
