+++ docs/rst/snippets/weighted_graph_14.rst
@@ -13,7 +13,7 @@ constant like so, adapting the path to your system:
 
 .. code:: python
 
-    CASE_PATH = r"C:\Users\myuser\git\ESA\tests\cases\ieee_14\IEEE 14 bus.pwb"
+    CASE_PATH = r"C:\Users\myuser\git\ESA\tests\cases\ieee_14\IEEE 14 bus_pws_version_21.pwb"
 
 Onward!
 
