+++ docs/rst/snippets/simple_graph_2000.rst
@@ -14,7 +14,7 @@ constant like so, adapting the path to your system:
 
 .. code:: python
 
-    CASE_PATH = r"C:\Users\myuser\git\ESA\tests\cases\tx2000\tx2000_base.PWB"
+    CASE_PATH = r"C:\Users\myuser\git\ESA\tests\cases\tx2000\tx2000_base_pws_version_21.pwb"
 
 On to the example!
 
