### 2020-05-29 a80c782
Adding note about Simulator versions to the prerequisites documentation. Also adding some initial notes to the changelog.
