+++ docs/rst/welcome/changelog.rst
@@ -2,6 +2,18 @@ Changes made with each ESA release are listed here. Please note that
 versions prior to 1.0.0 are not listed here, but are still available on
 `PyPi <https://pypi.org/project/esa/#history>`__.
 
+Version 1.0.3
+^^^^^^^^^^^^^
+
+* New SAW property, ``build_date``
+* New SAW helper method, ``get_simulator_version``
+* Updating so that ESA is compatible with Simulator version 17. Note
+  that this does not imply ESA has been tested with versions 16, 18, 19,
+  or 20. However, ESA *should* work with all these versions.
+* Added case files for Simulator versions 16-21 and renamed the cases
+  accordingly (suffixed with ``pws_version_<version goes here>.pwb``.
+* Updated documentation to discuss different versions of Simulator.
+
 Version 1.0.2
 ^^^^^^^^^^^^^
 
