+++ docs/rst/installation/pre_reqs.rst
@@ -8,6 +8,15 @@ ESA has the following prerequisites:
     the `Automation Server (SimAuto) add-on
     <https://www.powerworld.com/products/simulator/add-ons-2/simauto>`__
     installed.
+
+    * **NOTE**: the authors of ESA have tested with Simulator
+      versions 17 and 21. It is likely, but **not guaranteed**, that ESA
+      will work with all Simulator versions 16-21. If you encounter a
+      problem with a particular version, please file an `issue
+      <https://github.com/mzy2240/ESA/issues>`__ and we may be able
+      to help (if we can get access to that particular Simulator
+      version).
+
 *   `Git Large File Storage (LFS) <https://git-lfs.github.com/>`__
     (**OPTIONAL**: required to download case files and run tests). After
     installing Git LFS, simply change directories to the ESA repository,
