### 2020-05-29 99eeb13
1. Add support for version 22 and some beta versions.
