### 2020-05-29 c0a4f47
1. Rewrite the ProcessAuxFile test so that it will utilize a smaller case to test, which addresses https://github.com/mzy2240/ESA/issues/52
