+++ tests/constants.py
@@ -26,16 +26,13 @@ PATH_14_PWD = os.path.join(CASE_DIR, 'ieee_14', 'IEEE 14 bus.pwd')
 # Path to the Texas 2000 bus model.
 PATH_2000 = os.path.join(CASE_DIR, 'tx2000',
                          'tx2000_base_pws_version_{}.pwb'.format(VERSION))
-PATH_2000_mod = os.path.join(
-    CASE_DIR, 'tx2000_mod',
-    'ACTIVSg2000_AUG-09-2018_Ride_version7_pws_version_{}.pwb'.format(VERSION))
 
 # Path to the WSCC model.
 PATH_9 = os.path.join(CASE_DIR, 'wscc_9',
                       'WSCC 9 bus_pws_version_{}.pwb'.format(VERSION))
 
-# Aux file for filtering transformers by LTC control.
-LTC_AUX_FILE = os.path.join(THIS_DIR, 'ltc_filter.aux')
+# Aux file for filtering buses by area.
+AREA_AUX_FILE = os.path.join(THIS_DIR, 'area_filter.aux')
 
 # Map cases for doc testing.
 CASE_MAP = {'14': PATH_14, '2000': PATH_2000}
