### 2020-05-29 bf28e90
1. Remove the areanum from the default power flow fields, and ensure the test_snippets will pass.
