### 2020-05-29 286f61b
1. Run the coverage.
