+++ docs/rst/coverage.rst
@@ -1,11 +1,11 @@
-.. table:: ESA's testing coverage as of 2020-05-11 (Git commit: 9715a0a)
+.. table:: ESA's testing coverage as of 2020-05-29 (Git commit: bf28e90)
     :widths: auto
     :align: left
 
     +-----------------+-------------------+-----------------+-----------------+--------------------+
     | Name            |   Num. Statements |   Missing Lines |   Covered Lines |   Percent Coverage |
     +=================+===================+=================+=================+====================+
-    | esa/__init__.py |                 2 |               0 |               2 |                100 |
+    | esa/__init__.py |                 2 |               0 |               2 |           100      |
     +-----------------+-------------------+-----------------+-----------------+--------------------+
-    | esa/saw.py      |               332 |               0 |             332 |                100 |
+    | esa/saw.py      |               372 |              18 |             354 |            95.1613 |
     +-----------------+-------------------+-----------------+-----------------+--------------------+
