### 2020-05-29 bbae21a
Adding some tests so that running tests with the latest version of Simulator still achieves 100% coverage. This closes out https://github.com/mzy2240/ESA/issues/54
