### 2020-05-30 913d242
1. Update the builddate/version methodology and all the corresponding tests, which addresses https://github.com/mzy2240/ESA/issues/55
