### 2020-06-02 #2216 @danielskatz
👋 @tonyelhabr - would you be willing to review this submission for JOSS?  We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

In this particular case, you would also need access to PowerWorld and SimAuto.
