### 2020-06-04 #2289 @whedon
[REVIEW]: Easy SimAuto (ESA): A Python Package that Simplifies Interacting with PowerWorld Simulator
**Submitting author:** @blthayer (<a href="http://orcid.org/0000-0002-6517-1295">Brandon Thayer</a>)
**Repository:** <a href="https://github.com/mzy2240/ESA" target ="_blank">https://github.com/mzy2240/ESA</a>
**Version:** v1.0.6
