### 2020-06-10 #2289 @blthayer
@danielskatz - Thank you for kicking this off! I don't see `tonyelhabr` listed as an assignee on this issue - is that expected?

Also, I wanted to note that we tweaked the LaTex template so that a footnote appears indicating that myself and @mzy2240 contributed equally to the software. However, that footnote isn't showing up in the automatically generated .pdf file. What steps do we need to take to use our customized template before final publication? This isn't urgent since the review itself is still pending, but I figured it's best to bring it up earlier rather than later.
