### 2020-06-11 #2289 @danielskatz
> Thank you for kicking this off! I don't see tonyelhabr listed as an assignee on this issue - is that expected?

I think he will be marked as assigned once he accepts the invite as mentioned in the first comment in this issue.

>  What steps do we need to take to use our customized template before final publication?

@arfon - can you help here?
