### 2020-06-11 #2289 @arfon
Unfortunately that kind of custom modification isn't supported @blthayer. That said, something similar is possible using our existing template. I've downloaded your paper locally and changed the authors block to:

```
authors:
  - name: Brandon L. Thayer^[The first two authors of this paper contributed equally to this software.]
    orcid: 0000-0002-6517-1295
    affiliation: "1, 2"
  - name: Zeyu Mao
    orcid: 0000-0003-0841-5123
    affiliation: 1
  - name: Yijing Liu
    orcid: 0000-0002-5104-325X
    affiliation: 1
  - name: Katherine Davis
    orcid: 0000-0002-1603-1122
    affiliation: 1
  - name: Thomas Overbye
    orcid: 0000-0002-2382-2811
    affiliation: 1
```

...which results in the following PDF: [10.21105.joss.02289.pdf](https://github.com/openjournals/joss-reviews/files/4764477/10.21105.joss.02289.pdf)
