### 2020-06-11 8e47f62
We cannot customize the JOSS template. So, removing latex.template, updating README.rst, and updating paper.md to have the equal contribution footnote attached to Brandon's author field.
