### 2020-06-12 #2289 @danielskatz
👋 @tonyelhabr  - I see you've checked off all the items in your list, which I assume means you didn't find anything that needed to be improved - the submission meets all the required criteria.  Do you have any comments on the paper or software otherwise?
