### 2020-06-12 #2289 @blthayer
@danielskatz - I wanted to check if it's acceptable at this stage to swap out a single reference. Specifically, reference `(Thayer, 2020c)`. That reference refers to my Master's thesis, but the full text will not be publicly available until the Spring of 2022 (delayed release is conventional practice to avoid copyright issues when publishing portions of the thesis in academic journals). I've condensed my thesis into a journal article submission, and that submission should be available on arXiv starting Monday. So, I'd like to reference the arXiv journal article instead of the original thesis (which is currently unavailable except by request).
