### 2020-06-12 #2289 @tonyelhabr
@danielskatz I have no suggestions. I tested all examples in the documentation to verify software functionality. The examples are all non-trivial (with the exception of "Quick Start") and do bear resemblance to real-world usage (from my experience), which is good to see. The paper is concise and meets all of checklist items.
