### 2020-06-12 #2289 @blthayer
@danielskatz - thank you!
@tonyelhabr - thanks again for your review!

@tonyelhabr, @robinroche - please see [this comment](https://github.com/openjournals/joss-reviews/issues/2289#issuecomment-643388272) with regards to a very minor tweak I plan to make to a reference contained in the paper. I plan to make the reference change early next week (pending approval from arXiv).

I hope you all have great weekends!
