### 2020-06-12 #2289 @robinroche
@blthayer
I installed ESA without any issue and I am currently going through the quick start steps.
I get the following error:

```
(tutorial-env) C:\Users\rroche\Downloads>python test.py
22:55:06 [WARNING] [SAW]: UIVisible attribute could not be set. Note this SimAuto property was not introduced until Simulator version 20. Check your version with the get_simulator_version method.
Traceback (most recent call last):
  File "pandas\_libs\lib.pyx", line 1926, in pandas._libs.lib.maybe_convert_numeric
ValueError: Unable to parse string "  1,05999994"

Traceback (most recent call last):
  File "test.py", line 8, in <module>
    bus_data = saw.get_power_flow_results('bus')
  File "C:\Users\rroche\tutorial-env\lib\site-packages\esa\saw.py", line 458, in get_power_flow_results
    return self.GetParametersMultipleElement(ObjectType=object_type,
  File "C:\Users\rroche\tutorial-env\lib\site-packages\esa\saw.py", line 904, in GetParametersMultipleElement
    return self.clean_df_or_series(obj=df, ObjectType=ObjectType)
  File "C:\Users\rroche\tutorial-env\lib\site-packages\esa\saw.py", line 284, in clean_df_or_series
    obj[numeric_fields] = obj[numeric_fields].apply(pd.to_numeric)
  File "C:\Users\rroche\tutorial-env\lib\site-packages\pandas\core\frame.py", line 6878, in apply
    return op.get_result()
  File "C:\Users\rroche\tutorial-env\lib\site-packages\pandas\core\apply.py", line 186, in get_result
    return self.apply_standard()
  File "C:\Users\rroche\tutorial-env\lib\site-packages\pandas\core\apply.py", line 313, in apply_standard
    results, res_index = self.apply_series_generator()
  File "C:\Users\rroche\tutorial-env\lib\site-packages\pandas\core\apply.py", line 341, in apply_series_generator
    results[i] = self.f(v)
  File "C:\Users\rroche\tutorial-env\lib\site-packages\pandas\core\tools\numeric.py", line 149, in to_numeric
    values = lib.maybe_convert_numeric(
  File "pandas\_libs\lib.pyx", line 1963, in pandas._libs.lib.maybe_convert_numeric
ValueError: Unable to parse string "  1,05999994" at position 0
```
The problem appears to be with the last line of:
```
CASE_PATH = r"C:\Users\rroche\Downloads\IEEE 14 bus_pws_version_17.pwb"

from esa import SAW

saw = SAW(FileName=CASE_PATH)
saw.SolvePowerFlow()

bus_data = saw.get_power_flow_results('bus')
```
The warning seems normal as I am using PWS17.
For the parsing error, I suspect (but I may be wrong) that it may be related to the fact that I am using Windows 10 in French, for which the decimal point is actually a coma.
Running the same script without the last line returns no error, just the warning.
