### 2020-06-12 mzy2240/ESA#56 @blthayer
Add support for decimal delimiters other than periods
In Europe (and elsewhere), decimal numbers are delimited with a comma, rather than a period. Ensure that ESA seamlessly handles this difference.

This issue was brought up by `robinroche` in this [comment](https://github.com/openjournals/joss-reviews/issues/2289#issuecomment-643482550).
