### 2020-06-12 e6d0610
Automatically detecting PC's decimal delimiter, and adding a new _to_numeric method so that we can still use to_numeric from Pandas (which does not provide a decimal delimiter input). This should address https://github.com/mzy2240/ESA/issues/56#issue-638041265
