### 2020-06-12 661980d
Updating the changelog for the next release to mention automatic decimal delimiter detection.
