### 2020-06-12 #2289 @blthayer
@robinroche - thank you for bringing up this issue!

I agree - the warning you received is to be expected.

I have added support for comma-delimited decimal numbers. If you would like to get your hands on the changes right away, you can install ESA from source from our `develop` branch (found [here](https://github.com/mzy2240/ESA/tree/develop)).

Alternatively, I have submitted a pull request to be reviewed by the other ESA maintainers. After they've reviewed, a new ESA release will published including this fix. Regardless of which method you prefer, I will comment here when the next version is published and available to install via pip.

Thanks again, and have a great weekend!
