### 2020-06-14 7b330b8
1. Remove the HISTORY.rst file, which addresses https://github.com/mzy2240/ESA/issues/57
