+++ HISTORY.rst
@@ -1,7 +1,0 @@
-History
--------
-
-0.0.0
-+++++
-
-* Nothing
