+++ MANIFEST.in
@@ -1,1 +1,1 @@
-include README.rst HISTORY.rst LICENSE
+include README.rst docs/rst/welcome/changelog.rst LICENSE
