### 2020-06-14 #2289 @blthayer
@robinroche - the aforementioned changes have been released in ESA version `1.0.4`. I have noted the updates in the [changelog](https://mzy2240.github.io/ESA/html/welcome.html#changelog).

Assuming you installed ESA via pip, the process to upgrade is the same that you used for your initial installation, but you'll need to use pip's `--upgrade` flag. For example:

`python -m pip install --upgrade esa`

You can check the ESA version afterwards via:

```python
import esa
print(esa.__version__)
```

As I mentioned in my previous comment, please let me know if this fix doesn't work for you by commenting on [this issue](https://github.com/mzy2240/ESA/issues/56#issue-638041265). If the fix doesn't work for you, I'll work with you to collect more details so I can more directly reproduce the problem on my side and completely fix it.
