### 2020-06-14 mzy2240/ESA#56 @robinroche
@blthayer Thanks for the fix but I still have the same issue.
The update seems to have worked:
```
(tutorial-env) C:\Users\rroche>python -m pip install --upgrade esa
Collecting esa
  Downloading esa-1.0.4-py3-none-any.whl (46 kB)
     |████████████████████████████████| 46 kB 421 kB/s
Requirement already satisfied, skipping upgrade: pywin32 in c:\users\rroche\tutorial-env\lib\site-packages (from esa) (227)
Requirement already satisfied, skipping upgrade: pypiwin32 in c:\users\rroche\tutorial-env\lib\site-packages (from esa) (223)
Requirement already satisfied, skipping upgrade: pandas>=0.25 in c:\users\rroche\tutorial-env\lib\site-packages (from esa) (1.0.4)
Requirement already satisfied, skipping upgrade: numpy>=1.13.3 in c:\users\rroche\tutorial-env\lib\site-packages (from esa) (1.18.5)
Requirement already satisfied, skipping upgrade: pytz>=2017.2 in c:\users\rroche\tutorial-env\lib\site-packages (from pandas>=0.25->esa) (2020.1)
Requirement already satisfied, skipping upgrade: python-dateutil>=2.6.1 in c:\users\rroche\tutorial-env\lib\site-packages (from pandas>=0.25->esa) (2.8.1)
Requirement already satisfied, skipping upgrade: six>=1.5 in c:\users\rroche\tutorial-env\lib\site-packages (from python-dateutil>=2.6.1->pandas>=0.25->esa) (1.15.0)
Installing collected packages: esa
  Attempting uninstall: esa
    Found existing installation: esa 1.0.3
    Uninstalling esa-1.0.3:
      Successfully uninstalled esa-1.0.3
Successfully installed esa-1.0.4
```
I get the following error (file test.py has not changed since my first comment):
```
(tutorial-env) C:\Users\rroche\Downloads>python test.py
21:12:23 [WARNING] [SAW]: UIVisible attribute could not be set. Note this SimAuto property was not introduced until Simulator version 20. Check your version with the get_simulator_version method.
Traceback (most recent call last):
  File "pandas\_libs\lib.pyx", line 1926, in pandas._libs.lib.maybe_convert_numeric
ValueError: Unable to parse string "  1,05999994"

Traceback (most recent call last):
  File "test.py", line 8, in <module>
    bus_data = saw.get_power_flow_results('bus')
  File "C:\Users\rroche\tutorial-env\lib\site-packages\esa\saw.py", line 465, in get_power_flow_results
    return self.GetParametersMultipleElement(ObjectType=object_type,
  File "C:\Users\rroche\tutorial-env\lib\site-packages\esa\saw.py", line 912, in GetParametersMultipleElement
    return self.clean_df_or_series(obj=df, ObjectType=ObjectType)
  File "C:\Users\rroche\tutorial-env\lib\site-packages\esa\saw.py", line 290, in clean_df_or_series
    obj[numeric_fields] = self._to_numeric(obj[numeric_fields])
  File "C:\Users\rroche\tutorial-env\lib\site-packages\esa\saw.py", line 1859, in _to_numeric
    return data.apply(lambda x: pd.to_numeric(x, errors=errors))
  File "C:\Users\rroche\tutorial-env\lib\site-packages\pandas\core\frame.py", line 6878, in apply
    return op.get_result()
  File "C:\Users\rroche\tutorial-env\lib\site-packages\pandas\core\apply.py", line 186, in get_result
    return self.apply_standard()
  File "C:\Users\rroche\tutorial-env\lib\site-packages\pandas\core\apply.py", line 313, in apply_standard
    results, res_index = self.apply_series_generator()
  File "C:\Users\rroche\tutorial-env\lib\site-packages\pandas\core\apply.py", line 341, in apply_series_generator
    results[i] = self.f(v)
  File "C:\Users\rroche\tutorial-env\lib\site-packages\esa\saw.py", line 1859, in <lambda>
    return data.apply(lambda x: pd.to_numeric(x, errors=errors))
  File "C:\Users\rroche\tutorial-env\lib\site-packages\pandas\core\tools\numeric.py", line 149, in to_numeric
    values = lib.maybe_convert_numeric(
  File "pandas\_libs\lib.pyx", line 1963, in pandas._libs.lib.maybe_convert_numeric
ValueError: Unable to parse string "  1,05999994" at position 0
```
