### 2020-06-14 mzy2240/ESA#56 @blthayer
@robinroche - thank you very much for working with me on this, and I'm sorry the initial fix wasn't adequate.

So that I can better reproduce the issue, could you please execute the following (using the `CASE_PATH` variable for your own machine) and provide me with the verbatim printed output?

```python
from esa import SAW
from esa.saw import convert_list_to_variant

saw = SAW(FileName=CASE_PATH)

# Check decimal delimiter.
print('Decimal delimiter:')
print(saw.decimal_delimiter)
print('')

saw.SolvePowerFlow()

# Call SAW internal functions for debugging purposes.
output = saw._call_simauto(
    'GetParametersMultipleElement', 'bus',
    convert_list_to_variant(['BusNum', 'BusName', 'BusPUVolt', 'BusAngle',
                             'BusNetMW', 'BusNetMVR']), '')
print('Results from GetParametersMultipleElement:')
print(output)
print('')

```
