### 2020-06-14 mzy2240/ESA#56 @robinroche
@blthayer Sure, here is the output:

```
(tutorial-env) C:\Users\rroche\Downloads>python test2.py
21:54:21 [WARNING] [SAW]: UIVisible attribute could not be set. Note this SimAuto property was not introduced until Simulator version 20. Check your version with the get_simulator_version method.
Decimal delimiter:
.

Results from GetParametersMultipleElement:
(('    1', '    2', '    3', '    4', '    5', '    6', '    7', '    8', '    9', '   10', '   11', '   12', '   13', '   14'), ('Bus 1', 'Bus 2', 'Bus 3', 'Bus 4', 'Bus 5', 'Bus 6', 'Bus 7', 'Bus 8', 'Bus 9', 'Bus 10', 'Bus 11', 'Bus 12', 'Bus 13', 'Bus 14'), ('  1,05999994', '  1,04499996', '  1,00999999', '  1,01767162', '  1,01951459', '  1,07000005', '  1,06152019', '  1,09000003', '  1,05593281', '  1,05098559', '  1,05690694', '  1,05518907', '  1,05038265', '  1,03553058'), ('  0,00000000', ' -4,98255334', '-12,72502699', '-10,31282882', ' -8,77379918', '-14,22086891', '-13,35955842', '-13,35957101', '-14,93845750', '-15,09722071', '-14,79055172', '-15,07551240', '-15,15619553', '-16,03356498'), ('232,39169121', ' 18,30000132', '-94,19999719', '-47,79999852', ' -7,59999976', '-11,20000035', '  0,00000000', '  0,00000000', '-29,49999869', ' -9,00000036', ' -3,50000001', ' -6,10000007', '-13,50000054', '-14,90000039'), ('-16,54938906', ' 30,85595667', '  6,07485175', '  3,90000008', ' -1,60000008', '  5,22969961', '  0,00000000', ' 17,62306690', '  4,58488762', ' -5,79999983', ' -1,79999992', ' -1,60000008', ' -5,79999983', ' -5,00000007'))
```
