### 2020-06-14 mzy2240/ESA#56 @blthayer
@robinroche - thanks for your rapid response.

It's odd to me that it's reporting your decimal delimiter as a period, when it's clearly coming back from PowerWorld as a comma. This is my first time using Python's `locale` module, so I may be misunderstanding it.

Can you execute the following code and provide me with the output?

```python
import locale
print(locale.getdefaultlocale())
```
