### 2020-06-14 mzy2240/ESA#56 @robinroche
@blthayer Here is what I get:
```
(tutorial-env) C:\Users\rroche\Downloads>python
Python 3.8.3 (tags/v3.8.3:6f8c832, May 13 2020, 22:20:19) [MSC v.1925 32 bit (Intel)] on win32
>>> import locale
>>> print(locale.getdefaultlocale())
('fr_FR', 'cp1252')
```
