### 2020-06-14 mzy2240/ESA#56 @blthayer
@robinroche, that's very helpful, thank you. I _think_ I know how to fix this now.

To confirm, can you please try the following two things:
```python
import locale
locale.setlocale(locale.LC_NUMERIC, '')
d = locale.localeconv()
print(d['decimal_point'])
```
and
```python
import locale
locale.setlocale(locale.LC_NUMERIC, 'fr_FR')
d = locale.localeconv()
print(d['decimal_point'])
```
