### 2020-06-14 mzy2240/ESA#56 @robinroche
@blthayer Here is the output:
```
(tutorial-env) C:\Users\rroche\Downloads>python test3.py
,

(tutorial-env) C:\Users\rroche\Downloads>python test4.py
Traceback (most recent call last):
  File "test4.py", line 2, in <module>
    locale.setlocale(locale.LC_NUMERIC, 'fr_FR')
  File "C:\Users\rroche\AppData\Local\Programs\Python\Python38-32\lib\locale.py", line 608, in setlocale
    return _setlocale(category, locale)
locale.Error: unsupported locale setting
```
