### 2020-06-14 mzy2240/ESA#56 @blthayer
@robinroche - thank you. I am now quite sure I've figured out the correct way to fix this.

In my initial fix, I incorrectly assumed that the `locale` module automatically uses your current locale. In fact, you first have to set the locale, and the line `locale.setlocale(locale.LC_ALL, '')` sets all your locale information from your system.

I'll be able to push a fix to this right away, and hopefully @mzy2240 will be able to provide us with a new release for you to try shortly.
