### 2020-06-14 eb402a2
Adding code to the saw module to set the locale. This should finalize https://github.com/mzy2240/ESA/issues/56#issue-638041265. Also removing an unused import from test_saw.py.
