### 2020-06-14 1c04f33
Adding a test to make damn sure that our locale issues are fixed (at least from the perspective of calling `get_power_flow_results`).
