### 2020-06-14 #2289 @blthayer
@robinroche - we have a fix coming your way in ESA version `1.0.5`. I'll comment here when that's officially released.

If you would prefer to get started right away, adding the following to the _very top_ of your script should fix the issue for version `1.0.4`:

```python
import locale
locale.setlocale(locale.LC_ALL, '')
```
