### 2020-06-15 eb517bd
Fixing bug where POWER_FLOW_FIELDS['object type'] was inadvertently being extended in get_power_flow_results.
