+++ docs/rst/coverage.rst
@@ -1,4 +1,4 @@
-.. table:: ESA's testing coverage as of 2020-06-13 (Git commit: 661980d)
+.. table:: ESA's testing coverage as of 2020-06-15 (Git commit: eb517bd)
     :widths: auto
     :align: left
 
@@ -7,5 +7,5 @@
     +=================+===================+=================+=================+====================+
     | esa/__init__.py |                 2 |               0 |               2 |                100 |
     +-----------------+-------------------+-----------------+-----------------+--------------------+
-    | esa/saw.py      |               380 |               0 |             380 |                100 |
+    | esa/saw.py      |               382 |               0 |             382 |                100 |
     +-----------------+-------------------+-----------------+-----------------+--------------------+
