### 2020-06-15 #2289 @blthayer
@robinroche - excellent, thank you!

Also note that version `1.0.5` is now available if you'd like to upgrade. There's a very small bug fix in `1.0.5` in addition to adding the aforementioned `locale` lines to `saw.py`, but I don't think the bug affects any of the examples.
