### 2020-06-15 #2289 @blthayer
@robinroche - thank you! I'll file an issue in the ESA repository, and work on this issue this evening (Pacific time, West coast of USA).

I do suspect this is at least tangentially related to the locale issue, as I ran this example on my local machine and it works. So, I'll likely be giving you commands to execute so that I can get your output and reproduce on my end (like we did last time).

Thanks again for your time and review - your help in making our software work for older versions of PowerWorld and countries outside the US is **greatly** appreciated.

In the mean time, if you're able/willing to skip around to other examples so we can get any other issues knocked out as quickly as possible, I would appreciate it. Please note that one of the examples (the [weighted graph example](https://mzy2240.github.io/ESA/html/snippets.html#created-graph-model-with-edges-weighted-by-branch-impedance)) does not work with PWS 17, and the fact that it will not work with PWS 17 is noted in the docs.

Thanks!
