### 2020-06-15 mzy2240/ESA#61 @blthayer
Add Lines to Case example not working for all PWS versions or all locales
_Originally posted by @robinroche in https://github.com/openjournals/joss-reviews/issues/2289#issuecomment-644265540_

I'm now testing the "Add Lines to Case" example.
I'm getting the following:
```
(tutorial-env) C:\Users\rroche\Downloads>python test5.py
   From Number  To Number  Ckt        R        X        B  Lim MVA A
0         8155       5358    3  0.00037  0.00750  0.52342       2768
1         8154       8135    3  0.00895  0.03991  0.00585        149
2         8153       8108    3  0.01300  0.05400  0.02700        186
3         8152       8160    3  0.00538  0.03751  0.00613        221
4         8155       8057    3  0.00037  0.00750  0.52342       2768
5         8154       8153    3  0.01300  0.05400  0.02700        186
6         8155       8135    3  0.00538  0.03751  0.00613        221
19:18:30 [WARNING] [SAW]: UIVisible attribute could not be set. Note this SimAuto property was not introduced until Simulator version 20. Check your version with the get_simulator_version method.
Index(['BusNum', 'BusNum:1', 'LineCircuit', 'LineR', 'LineX', 'LineC',
       'LineAMVA'],
      dtype='object')
['BusNum', 'BusNum:1', 'LineCircuit']
Traceback (most recent call last):
  File "test5.py", line 34, in <module>
    saw.change_and_confirm_params_multiple_element('branch', line_df)
  File "C:\Users\rroche\tutorial-env\lib\site-packages\esa\saw.py", line 202, in change_and_confirm_params_multiple_element
    cleaned_df = self._change_parameters_multiple_element_df(
  File "C:\Users\rroche\tutorial-env\lib\site-packages\esa\saw.py", line 1754, in _change_parameters_multiple_element_df
    cleaned_df = self.clean_df_or_series(obj=command_df,
  File "C:\Users\rroche\tutorial-env\lib\site-packages\esa\saw.py", line 290, in clean_df_or_series
    obj[numeric_fields] = self._to_numeric(obj[numeric_fields])
  File "C:\Users\rroche\tutorial-env\lib\site-packages\esa\saw.py", line 1852, in _to_numeric
    data = data.apply(lambda x:
  File "C:\Users\rroche\tutorial-env\lib\site-packages\pandas\core\frame.py", line 6878, in apply
    return op.get_result()
  File "C:\Users\rroche\tutorial-env\lib\site-packages\pandas\core\apply.py", line 186, in get_result
    return self.apply_standard()
  File "C:\Users\rroche\tutorial-env\lib\site-packages\pandas\core\apply.py", line 295, in apply_standard
    result = libreduction.compute_reduction(
  File "pandas\_libs\reduction.pyx", line 620, in pandas._libs.reduction.compute_reduction
  File "pandas\_libs\reduction.pyx", line 128, in pandas._libs.reduction.Reducer.get_result
  File "C:\Users\rroche\tutorial-env\lib\site-packages\esa\saw.py", line 1853, in <lambda>
    x.str.replace(self.decimal_delimiter, '.'))
  File "C:\Users\rroche\tutorial-env\lib\site-packages\pandas\core\generic.py", line 5270, in __getattr__
    return object.__getattribute__(self, name)
  File "C:\Users\rroche\tutorial-env\lib\site-packages\pandas\core\accessor.py", line 187, in __get__
    accessor_obj = self._accessor(obj)
  File "C:\Users\rroche\tutorial-env\lib\site-packages\pandas\core\strings.py", line 2041, in __init__
    self._inferred_dtype = self._validate(data)
  File "C:\Users\rroche\tutorial-env\lib\site-packages\pandas\core\strings.py", line 2098, in _validate
    raise AttributeError("Can only use .str accessor with string values!")
AttributeError: Can only use .str accessor with string values!
```
Lines 33-34 in the script are:
```
(33) saw.RunScriptCommand("EnterMode(EDIT);")
(34) saw.change_and_confirm_params_multiple_element('branch', line_df)
```
I don't know if this is another locale issue, so just in case, I tried adding the following to the script, but the result remains the same:
```
import locale
locale.setlocale(locale.LC_ALL, '')
```
