### 2020-06-15 mzy2240/ESA#61 @blthayer
@mzy2240 - I'm busy today, so if you wanted to look into this, you're welcome to. Otherwise, I'll try to tackle it this evening. This could either be an issue with using PWS 17, or a locale issue (or something else entirely!). The example works for me with PWS 21.

We'll likely need to pinpoint where things are going wrong, and ask `robinroche` to run some internal `SAW` methods and give us the output so we can put it into a test case. See `test_get_power_flow_results_with_commas` for an example of how I've done this.

Also, please note that the aforementioned test method is not present on the master branch. So, something went wrong with your merging. Do you want me to file a separate ticket for that?
