### 2020-06-15 #2289 @robinroche
@blthayer So I tried all examples (see below). The good news is that only 2 examples are not working in my case, and the problem seems to be the same for both.

 - Install: OK
 - Quick start: OK (with 1.0.5)
 - Increase loading in case: error #1
 - Add lines to case: error #2
 - Create simple graph model: OK
 - Create graph model with edges weighted bu branch impedance: unable to test due to PWS17
 - Plot histogram of line flows with matplotlib: OK

 ERROR #1
 ```
 Traceback (most recent call last):
  File "test_increaseloading.py", line 16, in <module>
    saw.change_parameters_multiple_element_df('load', load_frame)
  File "C:\Users\rroche\tutorial-env\lib\site-packages\esa\saw.py", line 246, in change_parameters_multiple_element_df
    self._change_parameters_multiple_element_df(
  File "C:\Users\rroche\tutorial-env\lib\site-packages\esa\saw.py", line 1754, in _change_parameters_multiple_element_df
    cleaned_df = self.clean_df_or_series(obj=command_df,
  File "C:\Users\rroche\tutorial-env\lib\site-packages\esa\saw.py", line 290, in clean_df_or_series
    obj[numeric_fields] = self._to_numeric(obj[numeric_fields])
  File "C:\Users\rroche\tutorial-env\lib\site-packages\esa\saw.py", line 1852, in _to_numeric
    data = data.apply(lambda x:
  File "C:\Users\rroche\tutorial-env\lib\site-packages\pandas\core\frame.py", line 6878, in apply
    return op.get_result()
  File "C:\Users\rroche\tutorial-env\lib\site-packages\pandas\core\apply.py", line 186, in get_result
    return self.apply_standard()
  File "C:\Users\rroche\tutorial-env\lib\site-packages\pandas\core\apply.py", line 295, in apply_standard
    result = libreduction.compute_reduction(
  File "pandas\_libs\reduction.pyx", line 620, in pandas._libs.reduction.compute_reduction
  File "pandas\_libs\reduction.pyx", line 128, in pandas._libs.reduction.Reducer.get_result
  File "C:\Users\rroche\tutorial-env\lib\site-packages\esa\saw.py", line 1853, in <lambda>
    x.str.replace(self.decimal_delimiter, '.'))
  File "C:\Users\rroche\tutorial-env\lib\site-packages\pandas\core\generic.py", line 5270, in __getattr__
    return object.__getattribute__(self, name)
  File "C:\Users\rroche\tutorial-env\lib\site-packages\pandas\core\accessor.py", line 187, in __get__
    accessor_obj = self._accessor(obj)
  File "C:\Users\rroche\tutorial-env\lib\site-packages\pandas\core\strings.py", line 2041, in __init__
    self._inferred_dtype = self._validate(data)
  File "C:\Users\rroche\tutorial-env\lib\site-packages\pandas\core\strings.py", line 2098, in _validate
    raise AttributeError("Can only use .str accessor with string values!")
AttributeError: Can only use .str accessor with string values!
```

ERROR #2
```
Traceback (most recent call last):
  File "test_addlines.py", line 34, in <module>
    saw.change_and_confirm_params_multiple_element('branch', line_df)
  File "C:\Users\rroche\tutorial-env\lib\site-packages\esa\saw.py", line 202, in change_and_confirm_params_multiple_element
    cleaned_df = self._change_parameters_multiple_element_df(
  File "C:\Users\rroche\tutorial-env\lib\site-packages\esa\saw.py", line 1754, in _change_parameters_multiple_element_df
    cleaned_df = self.clean_df_or_series(obj=command_df,
  File "C:\Users\rroche\tutorial-env\lib\site-packages\esa\saw.py", line 290, in clean_df_or_series
    obj[numeric_fields] = self._to_numeric(obj[numeric_fields])
  File "C:\Users\rroche\tutorial-env\lib\site-packages\esa\saw.py", line 1852, in _to_numeric
    data = data.apply(lambda x:
  File "C:\Users\rroche\tutorial-env\lib\site-packages\pandas\core\frame.py", line 6878, in apply
    return op.get_result()
  File "C:\Users\rroche\tutorial-env\lib\site-packages\pandas\core\apply.py", line 186, in get_result
    return self.apply_standard()
  File "C:\Users\rroche\tutorial-env\lib\site-packages\pandas\core\apply.py", line 295, in apply_standard
    result = libreduction.compute_reduction(
  File "pandas\_libs\reduction.pyx", line 620, in pandas._libs.reduction.compute_reduction
  File "pandas\_libs\reduction.pyx", line 128, in pandas._libs.reduction.Reducer.get_result
  File "C:\Users\rroche\tutorial-env\lib\site-packages\esa\saw.py", line 1853, in <lambda>
    x.str.replace(self.decimal_delimiter, '.'))
  File "C:\Users\rroche\tutorial-env\lib\site-packages\pandas\core\generic.py", line 5270, in __getattr__
    return object.__getattribute__(self, name)
  File "C:\Users\rroche\tutorial-env\lib\site-packages\pandas\core\accessor.py", line 187, in __get__
    accessor_obj = self._accessor(obj)
  File "C:\Users\rroche\tutorial-env\lib\site-packages\pandas\core\strings.py", line 2041, in __init__
    self._inferred_dtype = self._validate(data)
  File "C:\Users\rroche\tutorial-env\lib\site-packages\pandas\core\strings.py", line 2098, in _validate
    raise AttributeError("Can only use .str accessor with string values!")
AttributeError: Can only use .str accessor with string values!
```
