### 2020-06-15 mzy2240/ESA#61 @blthayer
@mzy2240 - thank you! I think I have a hunch for how to fix the issue in this ticket. I'm happy to fix it myself this evening, or you're welcome to take it on.

In short, my new `_to_numeric` method assumes that all column data types are `object` (I think that's what Numpy/Pandas calls it). In other words, it assumes all columns are non-numeric.

To fix, we just need to adapt the `_to_numeric` method to check column data types in the beginning, and only operate on non-numeric columns.
