### 2020-06-15 d352471
1. Provide extra clarification to the TSGetContingencyResults usage.
