### 2020-06-16 3ed0340
Adding new unit test which replicates the issue in https://github.com/mzy2240/ESA/issues/61
