### 2020-06-16 393c021
Fixing _to_numeric so that in the event of both a comma decimal delimiter and data which is already numeric, we don't have problems. This wraps https://github.com/mzy2240/ESA/issues/61 and https://github.com/mzy2240/ESA/issues/56
