### 2020-06-16 d3b62b6
Updating paper.bib and paper.md to swap out Brandon's master's thesis with a journal paper that is pending submission, but has been posted to arXiv. This changed was discussed in https://github.com/openjournals/joss-reviews/issues/2289#issuecomment-643388272, https://github.com/openjournals/joss-reviews/issues/2289#issuecomment-643393981, and https://github.com/openjournals/joss-reviews/issues/2289#issuecomment-643467766.
