+++ docs/joss/paper.bib
@@ -39,13 +39,20 @@ title = {{PowerWorld Simulator}}
   ISSN={2378-3915},
   month={Sep.},}
 
-@mastersthesis{brandon_thesis,
-author = {Brandon Thayer},
-title = {Automating Voltage Control of the Electric Transmission Grid with {P}ower{W}orld and Deep Reinforcement Learning},
-school = {Texas A\&M University},
-year = 2020,
-month = {Mar},
-note = {Full text to be made publicly available in 2022.}
+@ARTICLE{brandon_arxiv,
+       author = {{Thayer}, Brandon L. and {Overbye}, Thomas J.},
+        title = "{Deep Reinforcement Learning for Electric Transmission Voltage Control}",
+      journal = {arXiv e-prints},
+     keywords = {Computer Science - Machine Learning, Electrical Engineering and Systems Science - Signal Processing, Electrical Engineering and Systems Science - Systems and Control, Statistics - Machine Learning},
+         year = 2020,
+        month = jun,
+          eid = {arXiv:2006.06728},
+        pages = {arXiv:2006.06728},
+archivePrefix = {arXiv},
+       eprint = {2006.06728},
+ primaryClass = {cs.LG},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020arXiv200606728T},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
 @misc{gym-powerworld,
