### 2020-06-16 624c642
Updating the changelog.
