+++ docs/rst/coverage.rst
@@ -1,4 +1,4 @@
-.. table:: ESA's testing coverage as of 2020-06-15 (Git commit: eb517bd)
+.. table:: ESA's testing coverage as of 2020-06-15 (Git commit: 624c642)
     :widths: auto
     :align: left
 
@@ -7,5 +7,5 @@
     +=================+===================+=================+=================+====================+
     | esa/__init__.py |                 2 |               0 |               2 |                100 |
     +-----------------+-------------------+-----------------+-----------------+--------------------+
-    | esa/saw.py      |               382 |               0 |             382 |                100 |
+    | esa/saw.py      |               387 |               0 |             387 |                100 |
     +-----------------+-------------------+-----------------+-----------------+--------------------+
