+++ README.rst
@@ -57,7 +57,7 @@ The ESA team works hard to ensure ESA is well tested, and we strive for
 testing coverage data for ESA, using `coverage
 <https://pypi.org/project/coverage/>`__.
 
-.. table:: ESA's testing coverage as of 2020-06-15 (Git commit: eb517bd)
+.. table:: ESA's testing coverage as of 2020-06-15 (Git commit: 624c642)
     :widths: auto
     :align: left
 
@@ -66,7 +66,7 @@ testing coverage data for ESA, using `coverage
     +=================+===================+=================+=================+====================+
     | esa/__init__.py |                 2 |               0 |               2 |                100 |
     +-----------------+-------------------+-----------------+-----------------+--------------------+
-    | esa/saw.py      |               382 |               0 |             382 |                100 |
+    | esa/saw.py      |               387 |               0 |             387 |                100 |
     +-----------------+-------------------+-----------------+-----------------+--------------------+
 
 License
