### 2020-06-16 #2289 @blthayer
@robinroche - thanks for your update, and thanks for the details.

We have a new release that _should_ address these issues: `v1.0.6`. If you could upgrade ESA and try those examples again, I'm hopeful that they'll work for you now.

Thanks again for your time and willingness to volunteer and help.
