### 2020-06-16 #2289 @robinroche
@blthayer I upgraded to 1.0.6 and all examples now work, thanks!

 - Install: OK
 - Quick start: OK
 - Increase loading in case: OK
 - Add lines to case: OK
 - Create simple graph model: OK
 - Create graph model with edges weighted by branch impedance: unable to test due to PWS17
 - Plot histogram of line flows with matplotlib: OK
