### 2020-06-16 #2289 @blthayer
@robinroche - excellent, I'm glad to hear it! Thanks for bearing with us as we addressed those locale issues.

As you go through the JOSS checklist, we'd love to hear further suggestions for improving ESA. Please don't hesitate to file issues over in our repository, even if it's as simple as "Section X of the documentation is confusing and should be revised."
