### 2020-06-16 #2289 @robinroche
@blthayer I have finished the review. Congratulations on this software, I wish it had existed when I needed it during my PhD! It is a much needed addition to PWS/SimAuto, simple to use and well documented.

Here are some minor improvement suggestions:
- In the documentation landing page, briefly explain what is SimAuto and what it can do (e.g., a few simple examples).
- Make a link to SimAuto's documentation more visible (I would expect most users would need it at some point).
- Clarify that the case path is the path to the .pwb file.
- State the goal of each example, and show results whenever possible (e.g., with power flow, but I acknowledge this may be difficult to keep it simple).
- List the main functions of SimAuto, so people can have an idea of what is possible with it and with ESA (with a link to SimAuto's documentation).

@danielskatz I have finished the review.
