### 2020-06-16 #2289 @blthayer
@robinroche - thank you! I appreciate your suggestions, and we'll get those worked in shortly.

@danielskatz, @mzy2240 - I'll plan to address @robinroche's comments tomorrow, either in the morning before work or in the evening afterwards.
