### 2020-06-17 071abec
Updating the .rst docs to address https://github.com/openjournals/joss-reviews/issues/2289#issuecomment-644975938
