### 2020-06-17 138ea0e
Building the documentation to address https://github.com/openjournals/joss-reviews/issues/2289#issuecomment-644975938
