+++ docs/html/_sources/index.rst.txt
@@ -6,6 +6,14 @@
 Easy SimAuto (ESA) Documentation
 ================================
 
+Easy SimAuto is a Python package to ease interaction with PowerWorld
+Simulator's :ref:`SimAuto <about-simauto>` add-on. This documentation
+seeks to be comprehensive, with detailed :ref:`installation instructions
+<installation>`, complete :ref:`API documentation <api>`, and a rich
+set of :ref:`examples <examples>`. If you're new to ESA, please start
+by reading the :ref:`welcome` section. Returning users looking to see
+what's new should check out the :ref:`changelog`.
+
 .. toctree::
    :maxdepth: 2
    :caption: Contents:
