### 2020-06-17 #2289 @blthayer
@robinroche, @danielskatz: I have addressed the documentation suggestions.

- Added new [About SimAuto](https://mzy2240.github.io/ESA/html/welcome.html#about-simauto) page and added text with links to the [table of contents/landing page](https://mzy2240.github.io/ESA/html/index.html)
- Links to SimAuto's documentation are available via the previously mentioned links
- All [examples](https://mzy2240.github.io/ESA/html/snippets.html) have language like `Before running the example below, define a CASE_PATH constant (the file path to a PowerWorld .pwb case file)` for clarification
- Any examples which did not already have their purpose stated up front now do
- All SimAuto functions and properties are listed in [About SimAuto](https://mzy2240.github.io/ESA/html/welcome.html#about-simauto)
