### 2020-06-17 #2289 @robinroche
@blthayer @danielskatz All suggestions have been addressed. This matches what I had in mind, thanks.
