+++ docs/joss/paper.bib
@@ -8,7 +8,7 @@ title = {{PowerWorld Simulator}}
 
 @misc{esa,
   author = {Brandon Thayer and Zeyu Mao and Yijing Liu},
-  title = {Easy SimAuto (ESA)},
+  title = {Easy Sim{A}uto ({ESA})},
   year = {2020},
   publisher = {GitHub},
   journal = {GitHub repository},
@@ -17,7 +17,7 @@ title = {{PowerWorld Simulator}}
 
 @misc{pypowerworld,
   author = {Jesse Mark Boyd},
-  title = {pypowerworld},
+  title = {{pypowerworld}},
   year = {2018},
   publisher = {GitHub},
   journal = {GitHub repository},
@@ -27,8 +27,8 @@ title = {{PowerWorld Simulator}}
 
 @INPROCEEDINGS{matpws,
   author={R. {Roche} and S. {Natarajan} and A. {Bhattacharyya} and S. {Suryanarayanan}},
-  booktitle={2012 23rd International Workshop on Database and Expert Systems Applications},
-  title={A Framework for Co-simulation of AI Tools with Power Systems Analysis Software},
+  booktitle={{2012 23rd International Workshop on Database and Expert Systems Applications}},
+  title={A Framework for Co-simulation of {AI} Tools with Power Systems Analysis Software},
   year={2012},
   volume={},
   number={},
@@ -57,7 +57,7 @@ archivePrefix = {arXiv},
 
 @misc{gym-powerworld,
   author = {Brandon Thayer},
-  title = {gym-powerworld},
+  title = {{gym-powerworld}},
   year = {2020},
   publisher = {GitHub},
   journal = {GitHub repository},
@@ -66,7 +66,7 @@ archivePrefix = {arXiv},
 
 @misc{drl-powerworld,
   author = {Brandon Thayer},
-  title = {drl-powerworld},
+  title = {{drl-powerworld}},
   year = {2020},
   publisher = {GitHub},
   journal = {GitHub repository},
@@ -75,7 +75,7 @@ archivePrefix = {arXiv},
 
 @INPROCEEDINGS{scenario_development,
   author={H. {Li} and J. H. {Yeo} and J. L. {Wert} and T. J. {Overbye}},
-  booktitle={2020 IEEE Texas Power and Energy Conference (TPEC)},
+  booktitle={{2020 IEEE Texas Power and Energy Conference (TPEC)}},
   title={Steady-State Scenario Development for Synthetic Transmission Systems},
   year={2020},
   volume={},
@@ -89,7 +89,7 @@ archivePrefix = {arXiv},
 
 @INPROCEEDINGS{OverbyeKPEC,
 author={A. Martinez and K. Garcia and C. Klauber and T. J. Overbye},
-booktitle={2020 Kansas Power and Energy Conference (KPEC)},
-title={Undergraduate Research on Design Considerations for a GMD Mitigation Systems},
+booktitle={{2020 Kansas Power and Energy Conference (KPEC)}},
+title={Undergraduate Research on Design Considerations for a {GMD} Mitigation Systems},
 year={2020},
 pages={1-6},}
