### 2020-06-17 mzy2240/ESA#64 @blthayer
Create final release for JOSS
@mzy2240 - could you please create a documentation-only release? No code has changed since `v1.0.6` was released, so I propose doing the following:

- Merge `master` into `develop` to incorporate commits from `danielskatz`
- Delete the `v1.0.6` tag both locally and remotely ([link](https://devconnected.com/how-to-delete-local-and-remote-tags-on-git/))
- Recreate the `v1.0.6` tag, pointing at the current tip of `master`
- Update `paper.pdf`? This is optional in my view.
- Do we need to replace the `v1.0.6` release on PyPi? I don't think those releases include the documentation, right? If the releases do not include the documentation, I don't think we need to do anything with respect to PyPi

Anything else? Please see this comment:
https://github.com/openjournals/joss-reviews/issues/2289#issuecomment-645538830

I'll then upload an archive of the software at v1.0.6 to Zenodo.
