### 2020-06-17 mzy2240/ESA#65 @blthayer
Upload v1.0.6 to Zenodo
See https://github.com/openjournals/joss-reviews/issues/2289#issuecomment-645538830
