### 2020-06-17 mzy2240/ESA#64 @blthayer
@mzy2240 - does the PyPi release have any of the documentation packaged with it? I don't think it does, but would need to double-check/investigate. If it doesn't, I don't think we need to change anything on the PyPi side.
