### 2020-06-17 mzy2240/ESA#64 @mzy2240
I just remove the old tag and re-create the new tag for v1.0.6.
Can you update the paper.pdf file? I think that is the only remaining part for this issue. Or just let me know where I can get the latest pdf file, I can update it also.
