### 2020-06-17 mzy2240/ESA#64 @blthayer
For this new tag, let's use the following (if you could upload it, I'd appreciate it):

https://github.com/openjournals/joss-reviews/issues/2289#issuecomment-645529505

After the paper is published, we'll update with the final version.
