### 2020-06-17 #2289 @blthayer
@danielskatz:

- tag; `v1.0.6
- Zenodo DOI: `10.5281/zenodo.3899462`
