### 2020-06-18 #2289 @blthayer
@danielskatz - thank you for your time and effort! Publishing in JOSS was a great experience, and I'll be paying it forward by signing up as a reviewer.
